Gate dependencies / binary behind features
A very minor change to shorten compile times. Now serde and serde_json
are only compiled when building our tests and clap and anyhow are only
compiled when building our binary.

The `gen-tests` was an existing feature, but now it's dependencies are
optional. A new feature named `build-binary` is used to contorl if a
binary is built.