Handle edge case with reference link definition titles
Markdown spec examples [209] and [210] help to illustrate that when
there's any text besides spaces, tabs, or newlines after something
that looks like a link reference definition title, either the entire
definition isn't a reference (209), or the quoted string that looks like
a title is actually part of the following paragraph (210).

This has been a long standing issue and I was ignoring these tests
because they were producing incorrect markdown. Now that I'm accounting
for this edge case these tests pass!

[209]: https://spec.commonmark.org/0.30/#example-209
[210]: https://spec.commonmark.org/0.30/#example-209