Apply patch to cmark-gfm/test/spec_tests.py and recover disabled tests
I just realized that some tests were not generated because they were
disabled. I'd like to generate all tests regardless if they were
disabled in the GitHub Flavored Markdown fork of commonmark, so I
applied a patch to the file and generated the missing tests.

For now I'm not going to vendor `test/spec_tests.py` and instead I'm
just going to hold onto the patch. The README in `./tests/spec/GitHub`
was updated to mention that the patch was applied.