Use the new link reference definition parser!
This improved the output of some test cases, and also made significant
changes to the codebase since the formatter is recovering these link
reference definitions instead of relying on `pulldown_cmark` to hand
them to it.

I feel like all the testing has helped to prove out the implementation,
though I wouldn't be shocked if there was an edge case that fails to
format nicely.