Add `CodeBlockContext` to code block formatter closure
The context provides info to help custom formatters correctly format the
code blocks within the bounds configured on the `FormatBuilder`.
For example, it provides access to the indentation and
`max_width` available to the code block. Of course custom formatters can
choose to ignore this information, but it's there if needed.