Add `spec_tests.py` and provide attribution
This script is used to parse spec files used for testing. The script
dumps a JSON output that this project uses to generate its own test
cases, and since I plan on adding more spec tests in the future I
thought it would be a good idea to have a local copy.