unify the various rewrite buffers into an enum
I found it a little cumbersome to add new rewrite buffers like footnote
definitions and headers because I'd need to add a new field to the
`FormatState` and then I'd also need to add methods to check if I was
currently formatting that markdown construct. Now we simply push the
`Writer` onto the stack, and we know that we should always write to that
top buffer. I think this simplifies the internals.