bump pulldown-cmark to v0.11.3
`pulldown-cmark v0.11.3` adds support for parsing math expressions and
adds tags to blockquote nodes. I'm not enabling these features because
it shouldn't be necessary to format these constructs. Math expressions
are written in paragraphs, and the new blockquote tags will just show up
as text that gets correctly formatted when it's there. I've added some
test cases to verify that math expressions and blockquotes are formatted
as expected.