fix a mistake in testing that prevented us from getting the test files
a while ago I changed the file extension in of the loose_list tests to
be `md` instead of `txt`. The caused the `check_adapter_events` test to
no longer pick up any files, and incorrectly report success when no work
was done.

After correcting this issue some events were also updated now that the
project is using a newer version of pulldown-cmark.