Add system for experimental configs and add unordered_list_marker config
I think the best way to add new unstable configs is behind a feature
flag. That way it's clear that the feature might be removed in the
future.

The new `unordered_list_marker` config allows users to normalize all
list markers to one of `*`, `+`, or `-` when enabled. Otherwise the
marker used in the source code will be inferred.