Add `WriteEvent` triat and `write_event!` and `writeln_event!` macros
This is a trait that helps pass additional context about the current
`Event` to the underlying `writer`. At the moment, this mostly helps the
Paragraph know which event is being written so it can make more informed
decisions on when to escape certain characters that look like other
markdown constructs.