use regex to match the delimiter row for a markdown table
I felt like I kept missing cases when it came to escaping text that
looked like a github flavored markdown tables so I decided to use a
regex for the match. I think this is a better approach, but I may
revisit how `could_be_table` is implemented in the future.