updated `needs_escape` to return an `EscapKind`
The `EscapKind` enum gives more context to the caller about what kind of
text the formatter thinks it's trying to escape. For now, I'm just
calling `.is_some()` in order to not mess with the current formatting
behavior, but the idea it to eventually start matching on the
`EscapeKind` to implement more context aware escaping.