update link reference definition tracing and minor optimization
The updated tracing helped me figure out that we could do things a
little more efficiently, and the improvement was to update the
`parsed_until` state to the end of the current input if we know there
are no more elements. This saves us from calling into
`parse_link_reference_definition` when we exhaust all the characters.

I think similar changes could be made when parsing titles, but I'm not
100% sure since I was recently updating the URL parsing so I just made
changes to that phase of the parser.