In nested contexts like blockquotes and lists indent code span content
Instead of adding backslash escapes to the code span, which don't
actually escape anything, I'm adding 4 spaces of indentation to those
lines. This technically also changes the content of the code span, but
when the line starts with a lazy continuation line the formatter adds
indentation that could change the way the markdown is parsed. By adding
spaces to these lines the parser continues to treat the text as a code
span. I believe those extra spaces will get removed when being rendered
as HTML. This also feels like a rare edge case so I don't know how often
users would experience this in practice.