No longer recover unknown links
recovering links from text that looked like link references was a real
source of idempotence issues so for now I've changed the implementation
never recover links. In the future I'd like to give users an option to
define the link recovery, but for now it'll stay disabled.

To keep the behavior of tests the same I've added the required link
reference definitions. In some cases the extra escaping I was doing was
no longer necessary so I was able to revert some ugly formatting, which
I think is a win.