hackey workaround for an odd parsing bug.
In one strange case, the parser will parse a definition list even when
it doesn't have a title. If that happens then we'll push a dummy escaped
title and some HTML comments questioning if this is what the user
actually wanted.

Also push HTML comments between the end of a definition list and a code
block to make sure we don't absorb a code block into the last definition
list definition.